+++ src/utilities/errors.test.ts
@@ -7,7 +7,6 @@ import {
   isInstantiationError,
   isUnknownTypeError,
   isValidationNoticesError,
-  safeStringify,
 } from './errors';
 
 describe('Errors Util', () => {
@@ -237,33 +236,5 @@ describe('Errors Util', () => {
         type: ErrorTypes.GLOBAL_SCHEDULING_CONDITIONS_FAILED,
       } as SchedulingError),
     ).deep.eq([1, 2]);
-  });
-});
-
-describe('safeStringify', () => {
-  test('Should stringify normal objects', () => {
-    const obj = { a: 1, b: 'test', c: true };
-    expect(safeStringify(obj)).toBe('{"a":1,"b":"test","c":true}');
-  });
-
-  test('Should handle circular references', () => {
-    const obj: Record<string, unknown> = { name: 'test' };
-    obj.self = obj;
-    const result = safeStringify(obj);
-    expect(result).toContain('[Circular]');
-  });
-
-  test('Should handle nested circular references', () => {
-    const parent: Record<string, unknown> = { name: 'parent' };
-    const child: Record<string, unknown> = { name: 'child', parent };
-    parent.child = child;
-    const result = safeStringify(parent);
-    expect(result).toContain('[Circular]');
-  });
-
-  test('Should fallback to toString for non-serializable values', () => {
-    const obj = { id: 123n };
-    const result = safeStringify(obj);
-    expect(result).toBe('[object Object]');
   });
 });
