+++ src/utilities/text.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from 'vitest';
-import { padNumber, pluralize, sanitizeCmdkValue } from './text';
+import { padNumber, pluralize, safeStringify, sanitizeCmdkValue } from './text';
 
 describe('sanitizeCmdkValue', () => {
   test('returns plain text unchanged', () => {
@@ -59,4 +59,32 @@ test('padNumber', () => {
   expect(padNumber(1, 3)).toBe('001');
   expect(padNumber(41, 5)).toBe('00041');
   expect(padNumber(10, 2)).toBe('10');
+});
+
+describe('safeStringify', () => {
+  test('Should stringify normal objects', () => {
+    const obj = { a: 1, b: 'test', c: true };
+    expect(safeStringify(obj)).toBe('{"a":1,"b":"test","c":true}');
+  });
+
+  test('Should handle circular references', () => {
+    const obj: Record<string, unknown> = { name: 'test' };
+    obj.self = obj;
+    const result = safeStringify(obj);
+    expect(result).toContain('[Circular]');
+  });
+
+  test('Should handle nested circular references', () => {
+    const parent: Record<string, unknown> = { name: 'parent' };
+    const child: Record<string, unknown> = { name: 'child', parent };
+    parent.child = child;
+    const result = safeStringify(parent);
+    expect(result).toContain('[Circular]');
+  });
+
+  test('Should fallback to toString for non-serializable values', () => {
+    const obj = { id: 123n };
+    const result = safeStringify(obj);
+    expect(result).toBe('[object Object]');
+  });
 });
